requirements.txt
pytest==8.3.5
allure-pytest==2.14.2
requests==2.32.3
allure-python-commons~=2.14.2
pydentic
jsonschema~=4.24.0
pydantic~=2.11.7